requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
Pillow==8.1.0
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
